requirements.txt
python-dotenv>=1.0.1
openai>=1.30.1
python-telegram-bot>=20.0
Flask>=2.3.3
httpx>=0.27.0
